Use environment variables instead of created variables
