+++ dev/View/Desktop/Interfaces/Windows/view/interfaces.cpp
@@ -1,5 +1,5 @@
 #include "../../system_interface.h"
-#ifdef IS_WIN32
+#ifdef WIN32
 #include <Windows.h>
 #include "window.h"
 #include "../view/view_interface.h"
