+++ dev/View/Desktop/Interfaces/Windows/view/view_interface.h
@@ -2,7 +2,7 @@
 #define VIEW_INTERFACE_PROC_H
 
 #include "../../system_interface.h"
-#ifdef IS_WIN32
+#ifdef WIN32
 
 #include <Windows.h>
 
@@ -12,6 +12,6 @@ extern HINSTANCE window_hInstance;
 
 LRESULT CALLBACK WindowProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
 
-#endif // IS_WIN32
+#endif // WIN32
 
 #endif // !VIEW_INTERFACE_PROC_H
